requirements.txt
anyio==3.6.2
boto3==1.26.86
botocore==1.29.86
bson==0.5.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dotenv-python==0.0.1
ecdsa==0.18.0
fastapi==0.93.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
loguru==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyMySQL==1.0.2
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
pytest==7.3.1
pytest-cov==4.1.0
